requirements.txt
msrest
azure-cognitiveservices-vision-face
django
djangorestframework
cryptography